General purpose command loader & Menu update
