+++ extension/src/extension.ts
@@ -16,8 +16,8 @@ import * as statusBar from "./statusbar";
 import * as autoIndent from './format/autoIndent';
 
 import * as DebugProviders from "./debug";
-import * as pmCmds from "./project/projectCommands";
-import * as helpCmds from './help/openWebHelp';
+import { registerProjectCommands } from "./project/projectCommands";
+import { registerCommands } from "./commands";
 import { loadAllResources } from "./resources";
 import * as nls from 'vscode-nls';
 
@@ -63,9 +63,9 @@ export function activate(context: vscode.ExtensionContext) {
 	});
 	
 	//-----------------------------------------------------------
-	//6. register project commands
-	pmCmds.registerProjectCommands(context);
-	helpCmds.registerHelpCommands(context);
+	//6. register commands
+	registerProjectCommands(context);
+	registerCommands(context);
 }
 
 export function deactivate(): Thenable<void> | undefined {
